+++ test/models/models.rb
@@ -9,6 +9,7 @@ class Person
   field :name,     type: String,  default: "me"
   field :age,      type: Integer
   field :email,    type: String
+  field :role,     type: String
 
   has_many :pets
 
@@ -17,6 +18,7 @@ class Person
   validates_uniqueness_of(:login, case_sensitive: false)
   validates_length_of(:password, minimum: 8, maximum: 16)
   validates_format_of(:email, with: /^([^\s]+)((?:[-a-z0-9]\.)[a-z]{2,})$/i)
+  validates_inclusion_of(:role, in: ["admin", "user"])
   validates_exclusion_of(:email, in: ["foo@bar.com", "fizz@buzz.com"])
 end